requirements/base.txt
aioredis==1.3.1
alabaster==0.7.12
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
Babel==2.8.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.1
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.5
django-appconf==1.0.4
django-betterforms==1.2
django-debug-toolbar==2.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.16
drf-nested-routers==0.91
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
incremental==17.5.0
Jinja2==2.11.2
livereload==2.6.1
markdown2==2.3.8
MarkupSafe==1.1.1
more-itertools==8.2.0
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
Pillow==7.1.1
pip-autoremove==0.9.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
python-magic==0.4.15
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
snowballstemmer==2.0.0
soupsieve==2.0
sqlparse==0.3.1
tornado==6.0.4
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
URLObject==2.4.3
whitenoise==5.0.1
zipp==3.1.0
zope.interface==5.1.0
